Fixed music widget
Fixed a host of issues with the music widget:

* Removed all functions related to fetching data from DOM. Doesn't work on  every server and causes issues with some server configs.
* Removed find_service_from_url(), it was replaced in the last version and there's no reason to keep it.
* Simplified find_service() to make it DRY.
* Simplified find_id() to go straight to finding the ID on page.
* Resolved bug that caused some players to display as "page cannot be found" - I had the parameter order wrong when I made the call to find the ID.
* Added albumortrack() to figure out if the URL is for an album, track, or is invalid

Removed functions.php - not needed.

Fixed the "hide this widget if empty" problem. It needed to be OR instead of AND in the check.

Updated Toolkit's find_on_page() to return FALSE if the page is empty. If that's not the case, then it continues on with the rest of the function instead of surrounding it in the if its not empty check.

Updated readme to with latest changes list.

Signed-off-by: Chris Wilcoxson <chris@slushman.com>